backend/requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
feedparser==6.0.10
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.2
Flask-CORS==4.0.0